past_case_retrieval/requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
sentence-transformers
pymupdf
numpy
python-multipart
pymongo